Defeat USB Malware
This registry tweak tells Windows to treat AUTORUN.INF as if it were a configuration file from a pre-Windows 95 application. This means it says "whenever you have to handle a file called AUTORUN.INF, don't use the values from the file. You'll find alternative values at HKEY_LOCAL_MACHINE\SOFTWARE\DoesNotExist." And since that key does not exist, it will act like as if AUTORUN.INF is completely empty, and so nothing autoruns, and nothing is added to the Explorer double-click action. The benfit is that it works even if yourned of UAC (like some people sadly still do). 

So no Malware can hit you unless you explicitly click the executable file.