Microsoft Security Servicing Criteria for Windows added
Officially explanation when MS patches bugs and when not. E.g. UAC bypasses are not been patched cause it's one of the 15 "Defense-in-depth Security Features" from Windows such as AppLocker, ASLR etc. However, you get a reward (Security Boundaries) for bypassing 7 mechanisms. 

The reason why is that just because there is a bug or hole doesn't mean it can automatically be bypassed or abused (Social Engineering + Bug abuse). Ms only closes such bigger holes with security feature updates - NOT a patchday.